Add info parameter to driver choice list
Add a parameter to driver choice list which allows the method to be reused in cases where excluded drivers names need to be hidden or shown. For example, the driver info command requires all drivers in the championship to be listed, while the draft command needs only to show active drivers.